CORE-1299 Move call for running vice jobs to hook
